Added checks on userId in MealRepository's implementation

+++ src/main/java/ru/javawebinar/topjava/repository/inmemory/InMemoryMealRepository.java
@@ -1,10 +1,13 @@
 package ru.javawebinar.topjava.repository.inmemory;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Repository;
 import ru.javawebinar.topjava.model.Meal;
 import ru.javawebinar.topjava.repository.MealRepository;
 import ru.javawebinar.topjava.util.DateTimeUtil;
 import ru.javawebinar.topjava.util.MealsUtil;
+import ru.javawebinar.topjava.web.SecurityUtil;
 
 import java.time.LocalDate;
 import java.util.ArrayList;
@@ -17,47 +20,60 @@ import java.util.stream.Collectors;
 
 @Repository
 public class InMemoryMealRepository implements MealRepository {
+    private static final Logger log = LoggerFactory.getLogger(InMemoryMealRepository.class);
     private Map<Integer, Meal> repository = new ConcurrentHashMap<>();
     private AtomicInteger counter = new AtomicInteger(0);
 
-//    {
-//        for (Meal MEAL : MealsUtil.MEALS) {
-//            save(MEAL, SecurityUtil.authUserId());
-//        }
-//    }
-
+    {
+        for (Meal MEAL : MealsUtil.MEALS) {
+            save(MEAL, SecurityUtil.authUserId());
+        }
+    }
 
     @Override
     public Meal save(Meal meal, int userId) {
+        log.info("save {}", meal);
         if (meal.isNew()) {
             meal.setId(counter.incrementAndGet());
             repository.put(meal.getId(), meal);
             return meal;
         }
-        // handle case: update, but not present in storage
-        return repository.computeIfPresent(meal.getId(), (id, oldMeal) -> meal);
+
+        if (userId != meal.getUserId()) {
+            return null;
+        } else {
+            // handle case: update, but not present in storage
+            return repository.computeIfPresent(meal.getId(), (id, oldMeal) -> meal);
+        }
     }
 
     @Override
     public boolean delete(int id, int userId) {
-        return repository.remove(id) != null;
+        log.info("delete {}", id);
+        if (repository.get(id).getUserId() == userId) {
+            return repository.remove(id) != null;
+        } else return false;
     }
 
     @Override
     public Meal get(int id, int userId) {
-        return repository.get(id);
+        log.info("get {}", id);
+        Meal meal = repository.get(id);
+        return (meal.getUserId() == userId ? meal : null);
     }
 
     @Override
     public List<Meal> getAll(int userId) {
+        log.info("getAll");
         List<Meal> mealList = repository.values().stream()
-                .sorted(Comparator.comparing(Meal::getDate))
+                .sorted(Comparator.comparing(Meal::getDate).reversed())
                 .collect(Collectors.toList());
-        return MealsUtil.filterByUserId(mealList, userId);
+        return MealsUtil.filterListByUserId(mealList, userId);
     }
 
     @Override
     public List<Meal> getBetween(LocalDate startDate, LocalDate endDate, int userId) {
+        log.info("get between {} and {}", startDate, endDate);
         List<Meal> mealList = new ArrayList<>();
         for (Meal meal : repository.values()) {
             if (meal.getUserId() == userId) {
